+++ src/app/(route)/mypage/inquiries/_components/MypageInquiriesContent/MypageInquiriesContent.tsx
@@ -4,7 +4,7 @@ import { InquiryItemType, useGetUserInquiries } from "@/api/fetch/inquiry";
 import { Chip, MypageEmptyUI, LoadingState } from "@/components";
 import { useToast } from "@/context/ToastContext";
 import { useInfiniteScroll } from "@/hooks";
-import { useFilterParams } from "@/hooks/domain";
+import { useFilterParams } from "@/hooks";
 import { formatDate, highlightText } from "@/utils";
 import Link from "next/link";
 import { useEffect } from "react";
