+++ src/app/(route)/mypage/reports/_components/MypageReportsFilter/MypageReportsFilter.test.tsx
@@ -4,7 +4,7 @@ import MypageReportsFilter from "./MypageReportsFilter";
 const mockUpdateFilters = jest.fn();
 const mockUseFilterParams = jest.fn();
 
-jest.mock("@/hooks/domain", () => ({
+jest.mock("@/hooks", () => ({
   useFilterParams: () => mockUseFilterParams(),
   useFilterSync: () => ({ updateFilters: mockUpdateFilters }),
 }));
