+++ src/app/(route)/mypage/reports/_components/MypageReportsFilter/MypageReportsFilter.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Filter, KebabMenu } from "@/components";
-import { useFilterParams, useFilterSync } from "@/hooks/domain";
+import { useFilterParams, useFilterSync } from "@/hooks";
 import { filterSelectionState, normalizedFilterValues, normalizeEnumValue } from "@/utils";
 import { useState } from "react";
 import {
