+++ src/app/(route)/sign-up/_hooks/useSignUpBtnClick.ts
@@ -4,7 +4,7 @@ import { useEffect, useMemo, useState } from "react";
 import { EMAIL_ERROR_MESSAGE, EMAIL_CHECK_CODE_MESSAGE } from "../_constants/SIGNUP_ERROR_MESSAGE";
 import { throttle } from "lodash";
 import { useApiCheckCode, useApiSendEmail } from "@/api/fetch/auth";
-import { useErrorToast, useNicknameCheck } from "@/hooks/domain";
+import { useErrorToast, useNicknameCheck } from "@/hooks";
 import { FormType } from "../types/FormType";
 
 type SignUpFieldType = Omit<FormType, "privacyPolicyAgreed" | "marketingConsent">;
