+++ src/app/(route)/sign-up/_hooks/useSignUpSubmit.test.ts
@@ -17,7 +17,7 @@ jest.mock("@/context/ToastContext", () => ({
   useToast: () => ({ addToast: mockAddToast }),
 }));
 
-jest.mock("@/hooks/domain", () => ({
+jest.mock("@/hooks", () => ({
   useErrorToast: () => ({ handlerApiError: mockHandlerApiError }),
 }));
