+++ src/app/(route)/sign-up/_hooks/useSignUpSubmit.ts
@@ -2,7 +2,7 @@ import { useRouter } from "next/navigation";
 import { useToast } from "@/context/ToastContext";
 import { ApiSignUpType, useApiSignUp } from "@/api/fetch/auth";
 import { SIGNUP_ERROR_MESSAGE } from "../_constants/SIGNUP_ERROR_MESSAGE";
-import { useErrorToast } from "@/hooks/domain";
+import { useErrorToast } from "@/hooks";
 import { usePermissionStore } from "@/store";
 import { useQueryClient } from "@tanstack/react-query";
 import { throttle } from "lodash";
